pyproject.toml
[tool.poetry]
name = "cosmos"
version = "0.1.0"
description = ""
authors = ["YeonwooSung <neos960518@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.99.1"
slowapi = "^0.1.8"
uvicorn = "^0.22.0"
apache-airflow = "^2.6.2"
starlette-admin = "^0.10.0"
parse = "^1.19.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
